backend: protecting API routes, frontend: protecting React routes, making authorized requests

+++ frontend/src/components/WorkoutDetails.js
@@ -3,17 +3,23 @@ import React from 'react';
 import { baseUrl } from '../shared/baseUrl';
 // hooks
 import { useWorkoutsContext } from '../hooks/useWorkoutsContext';
+import { useAuthContext } from '../hooks/useAuthContext';
 // date fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow';
 
 export default function WorkoutDetails({workout}) {
 
   const {dispatch} = useWorkoutsContext();
+  const { user } = useAuthContext();
 
   const handleClick = async () => {
+    if (!user) return
 
     const response = await fetch(`${baseUrl}/api/workouts/${workout._id}`, {
-      method: 'DELETE'
+      method: 'DELETE',
+      headers: {
+        'Authorization': `Bearer ${user.token}`
+      }
     });
     const datas = await response.json();
